Update README.md
opms-online product management system, Is continuous project  which allow user to order different product from different area. This is not complete project i m still deal with this project in order to accomplish it.
On this project Admin can see all order that user can offer and verify those order as well as admin can able to add product when he/she think it is small for a certain order.